Fix duplicate ids when saving wf with inline steps
Instead of the file name (which is duplicated when a step is added to
the workflow multiple times), the name of the step in the workflow is
used for id fields of inline steps. The name of the step in the workflow
is unique.

Also added a test to test this.

Fixes #82