Add option to save workflow with --pack option
This is an alternative to saving steps inline (which doesn't always work).

Packed workflows cannot be loaded into scriptcwl anymore (so, they can't
be used as building blocks). Added a test to make sure the saving as packed
works and the resulting workflow is not loaded into the steps library.