Fix saving of relative paths for workflows with steps from urls
If the created workflow contained a step loaded from a url and the
workflow was saved with relative paths, the relative path was
concatinated to the relative path. This resulted in a validation
error. The error was fixed by checking if the step was loaded from
a url. Also added a test for this situation.