Add option to use working_dir for saving workflows
When creating a workflow using the WorkflowGenerator object, the user
can specify a working directory. All loaded cwl files are copied to
the working directory, and when saving a workflow with the wd option,
the run field of steps in a workflow consist of the basename of a step
(this works, because a copy of each cwl file is in the working directory).

Todo:

- remove repeated parts of code
- add tests
- add documentation