Load tools before workflows
If a working directory is used, tools must be loaded in the steps
library before workflows, otherwise some workflows may not be loaded
because steps are missing (i.e. they have not yet been copied to
the working directory).

If a workflow is used as a subworkflow it must be loaded before the
workflow it is a subworkflow in. This problem is not yet fixed.

Refs #94