Fix adding of scatter_method=None
The scatter_method keyword is optional when scattering over a single
parameter (#103). This was fixed before. However, when saving the
workflow, the step's scatter_method was set to None, which lead to
a validation error when saving the workflow. This is fixed by checking
whether the scatter_method is set (and should therefore be added).

Also updated the test to validate the workflow after adding a scattered
step without scatter_method.